Update README.md
The notebook focuses on building a deep learning model to classify different types of meat using TensorFlow and Keras. It imports necessary libraries, sets paths for training and testing datasets, and resizes images to a consistent 800x800 pixel size. The dataset is processed into TensorFlow-compatible format, shuffled, and batched for efficient training. A ResNet50-based model is constructed with custom layers for classification and compiled with the Adam optimizer. The model is trained using early stopping and learning rate reduction for optimal performance. After training, it is evaluated on a test dataset, generating metrics like accuracy, precision, recall, and a confusion matrix. Finally, the model is saved for future use.